Fix cross-platform compilation for darwin and windows
Split unix.TCGETS terminal detection into color_linux.go/color_other.go
so darwin/windows builds don't pull in the linux-only unix package.
Add platform_other.go stub so newLinuxPlatform symbol resolves on all targets.
