+++ idspoof/internal/ui/color.go
@@ -1,21 +1,6 @@
 package ui
 
-import (
-	"os"
-
-	"golang.org/x/sys/unix"
-)
-
 var colorsEnabled bool
-
-func init() {
-	colorsEnabled = isTerminal(os.Stdout.Fd())
-}
-
-func isTerminal(fd uintptr) bool {
-	_, err := unix.IoctlGetTermios(int(fd), unix.TCGETS)
-	return err == nil
-}
 
 // ANSI escape codes.
 const (
